refactor(adapters/langgraph): apply review feedback on the output guard
- dunder-wrap the guard's state key (__structured_output_attempts__) so
  it cannot collide with a field the agent's own state schema declares
- document how after_model classifies each turn and why an
  unclassifiable one leaves the counter untouched instead of resetting
- rename _structured_output.py to _agent_output_guard.py and the test
  file to match; the is_single_string_output test moves next to _utils
